+++ src/main/java/DBTranslate/logic/MySqlService.java
@@ -84,12 +84,12 @@ public class MySqlService implements IMySqlService {
 	        String sql = "Select * from LOCATIONS";
 	        ResultSet resultSet = statment.executeQuery(sql);
 			while(resultSet.next()){
-				 long location_id  = resultSet.getLong(resultSet.getMetaData().getColumnName(1));
+				 int location_id  = resultSet.getInt(resultSet.getMetaData().getColumnName(1));
 				 String street_address = resultSet.getString(resultSet.getMetaData().getColumnName(2));
 				 int  postal_code = resultSet.getInt(resultSet.getMetaData().getColumnName(3));
 				 String city = resultSet.getString(resultSet.getMetaData().getColumnName(4)); 
 				 String state_province = resultSet.getString(resultSet.getMetaData().getColumnName(5));
-				 long  country_id = resultSet.getLong(resultSet.getMetaData().getColumnName(6));
+				 int  country_id = resultSet.getInt(resultSet.getMetaData().getColumnName(6));
 			     listOfISqlTableDTO.add(new LocationsSqlTableDTO(location_id, street_address, postal_code, city, state_province, country_id));
 			}
 			ISqlTableDTO[] allData = new ISqlTableDTO[listOfISqlTableDTO.size()];
@@ -225,17 +225,17 @@ public class MySqlService implements IMySqlService {
 	        String sql = "Select * from EMPLOYEES";
 	        ResultSet res = stmt.executeQuery(sql);
 			while(res.next()) {
-				 long employee_id =  res.getLong(res.getMetaData().getColumnName(1));
+				 int employee_id =  res.getInt(res.getMetaData().getColumnName(1));
 				 String first_name = res.getString(res.getMetaData().getColumnName(2));
 				 String last_name= res.getString(res.getMetaData().getColumnName(3));
 				 String email = res.getString(res.getMetaData().getColumnName(4));
 				 String phone_number = res.getString(res.getMetaData().getColumnName(5));
 				 Date hire_date  = res.getDate(res.getMetaData().getColumnName(6));
-				 long job_id  =  res.getLong(res.getMetaData().getColumnName(7));
+				 int job_id  =  res.getInt(res.getMetaData().getColumnName(7));
 				 int salary  =  res.getInt(res.getMetaData().getColumnName(8));
 				 String commission_pct  = res.getString(res.getMetaData().getColumnName(9));
-				 long manager_id =  res.getLong(res.getMetaData().getColumnName(10));
-				 long department_id=  res.getLong(res.getMetaData().getColumnName(11));
+				 int manager_id =  res.getInt(res.getMetaData().getColumnName(10));
+				 int department_id=  res.getInt(res.getMetaData().getColumnName(11));
 				 listOfISqlTableDTO.add(new EmployeesSqlTableDTO(employee_id, first_name, last_name, email, phone_number, hire_date, job_id, salary, commission_pct, manager_id, department_id));
 			}
 			ISqlTableDTO[] data = new ISqlTableDTO[listOfISqlTableDTO.size()];
